resources completely refactored
