+++ src/main/java/com/tribesbackend/tribes/tribeskingdom/model/Kingdom.java
@@ -2,7 +2,7 @@ package com.tribesbackend.tribes.tribeskingdom.model;
 
 
 import com.tribesbackend.tribes.tribesbuilding.model.Building;
-import com.tribesbackend.tribes.tribesresource.model.ResourcesModel;
+import com.tribesbackend.tribes.tribesresources.model.ResourcesModel;
 import com.tribesbackend.tribes.tribesuser.model.TribesUser;
 import com.tribesbackend.tribes.troop.model.Troop;
