--- src/main/java/com/tribesbackend/tribes/tribesresource/model/ResourcesModel.java
@@ -1,82 +1,0 @@
-package com.tribesbackend.tribes.tribesresource.model;
-
-import com.tribesbackend.tribes.tribeskingdom.model.Kingdom;
-import org.springframework.validation.annotation.Validated;
-
-import javax.persistence.*;
-import javax.validation.constraints.Min;
-import javax.validation.constraints.NotNull;
-import java.sql.Timestamp;
-
-
-@Validated
-@Entity
-@Table(name = "resourcesModels")
-public class ResourcesModel {
-    @Id
-    @GeneratedValue
-    public long resourcesId;
-
-    public String type;
-
-    @Min(value = 0L, message = "The value must be positive")
-    @NotNull
-    long amount;
-
-    long timeStampLastVisit;
-
-    @ManyToOne(fetch = FetchType.LAZY, optional = false)
-    @JoinColumn(name = "kingdom_id", nullable = false)
-    Kingdom kingdom;
-
-    public ResourcesModel() {
-    }
-
-    public ResourcesModel(String type) {
-        this.type = type;
-    }
-
-    /*public ResourcesModel(String type, Kingdom kingdom) {
-        if (!type.equals("gold")||!type.equals("food")){
-            throw new IllegalArgumentException();
-        }
-        else this.type = type;
-
-        switch(type){
-            case "gold":
-                this.amount = 100;// to be specified
-                break;
-            case "food":
-                this.amount = 0;
-                break;
-        }
-        this.kingdom = kingdom;
-    }*/
-
-    public String getType() {
-        return type;
-    }
-
-    public void setType(String type) {
-        if (!type.equals("gold")||!type.equals("food")){
-            throw new IllegalArgumentException();
-        }
-        else this.type = type;
-    }
-
-    public long getAmount() {
-        return amount;
-    }
-
-    public void setAmount(long amount) {
-        this.amount = amount;
-    }
-
-    public long getTimeStampLastVisit() {
-        return timeStampLastVisit;
-    }
-
-    public void setTimeStampLastVisit(long timeStampLastVisit) {
-        this.timeStampLastVisit = timeStampLastVisit;
-    }
-}
