--- src/main/java/com/tribesbackend/tribes/tribesresource/repository/ResourceRepository.java
@@ -1,18 +1,0 @@
-package com.tribesbackend.tribes.tribesresource.repository;
-
-import com.tribesbackend.tribes.tribesresource.model.ResourcesModel;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-import java.util.Optional;
-
-@Repository
-public interface ResourceRepository extends JpaRepository <ResourcesModel, Long> {
-    @Override
-    List<ResourcesModel> findAll();
-
-    Optional<ResourcesModel> findResourceByResourcesId(long id);
-
-
-}
