+++ src/main/java/com/tribesbackend/tribes/tribesresources/model/ResourceService.java
@@ -1,7 +1,7 @@
-package com.tribesbackend.tribes.tribesresource.model;
+package com.tribesbackend.tribes.tribesresources.model;
 
 import com.tribesbackend.tribes.service.TimeService;
-import com.tribesbackend.tribes.tribesresource.repository.ResourceRepository;
+import com.tribesbackend.tribes.tribesresources.repository.ResourceRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
