+++ src/main/java/com/tribesbackend/tribes/tribesresources/model/ResourcesFactory.java
@@ -1,5 +1,7 @@
 package com.tribesbackend.tribes.tribesresources.model;
 
+package com.tribesbackend.tribes.tribesresources.model;
+
 public class ResourcesFactory {
     public static Resources createValidSampleResources() {
         return new Resources.ResourcesBuilder()
