+++ src/main/java/com/tribesbackend/tribes/tribesresources/repository/ResourceRepository.java
@@ -1,13 +1,18 @@
 package com.tribesbackend.tribes.tribesresources.repository;
 
-import com.tribesbackend.tribes.tribesresources.model.Resources;
+import com.tribesbackend.tribes.tribesresources.model.ResourcesModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
+import java.util.Optional;
 
 @Repository
-public interface ResourceRepository extends JpaRepository<Resources, Long> {
-    List<Resources> findAll();
+public interface ResourceRepository extends JpaRepository <ResourcesModel, Long> {
+    @Override
+    List<ResourcesModel> findAll();
+
+    Optional<ResourcesModel> findResourceByResourcesId(long id);
+
 
 }
